--- src/Equinor.ProCoSys.BusReceiver/Interfaces/IBusReceiverServiceFactory.cs
@@ -1,7 +1,0 @@
-﻿namespace Equinor.ProCoSys.BusReceiver.Interfaces
-{
-    public interface IBusReceiverServiceFactory
-    {
-        IBusReceiverService GetServiceInstance();
-    }
-}
