--- src/Equinor.ProCoSys.BusReceiver/Interfaces/IPcsSubscriptionClient.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Azure.ServiceBus;
-
-namespace Equinor.ProCoSys.BusReceiver.Interfaces
-{
-    public interface IPcsSubscriptionClient
-    {
-        PcsTopic PcsTopic { get; }
-        void RegisterPcsMessageHandler(Func<IPcsSubscriptionClient, Message, CancellationToken, Task> handler, MessageHandlerOptions messageHandlerOptions);
-        Task CompleteAsync(string token);
-        Task CloseAsync();
-    }
-}
