--- src/Equinor.ProCoSys.BusReceiver/Interfaces/IPcsSubscriptionClients.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Azure.ServiceBus;
-
-namespace Equinor.ProCoSys.BusReceiver.Interfaces
-{
-    public interface IPcsSubscriptionClients
-    {
-        Task CloseAllAsync();
-        void RegisterPcsMessageHandler(Func<IPcsSubscriptionClient, Message, CancellationToken, Task> handler, MessageHandlerOptions options);
-    }
-}
