+++ src/Equinor.ProCoSys.BusReceiver/Receiver/Interfaces/IBusReceiverService.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Azure.ServiceBus;
+
+namespace Equinor.ProCoSys.BusReceiver.Receiver.Interfaces
+{
+    public interface IBusReceiverService
+    {
+        Task ProcessMessageAsync(PcsTopic pcsTopic, Message message, CancellationToken token);
+    }
+}
