+++ src/Equinor.ProCoSys.BusReceiver/Receiver/Interfaces/IBusReceiverServiceFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace Equinor.ProCoSys.BusReceiver.Receiver.Interfaces
+{
+    public interface IBusReceiverServiceFactory
+    {
+        IBusReceiverService GetServiceInstance();
+    }
+}
