+++ src/tests/Equinor.ProCoSys.BusReceiverTests/Equinor.ProCoSys.BusReceiverTests.csproj
@@ -12,7 +12,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Equinor.ProCoSys.BusReceiver\Equinor.ProCoSys.BusReceiver.csproj" />
+    <ProjectReference Include="..\..\Equinor.ProCoSys.BusReceiver\Equinor.ProCoSys.PcsBus.csproj" />
   </ItemGroup>
 
 </Project>
